requirements.txt
Pillow
torch
torchvision
